# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: RentFlow API
  description: |
    Multi-tenant Rent Management SaaS API.

    **Rules enforced at every endpoint:**
    - `public_id` is used in all URLs and responses — internal `id` is never exposed
    - All resources are scoped to the authenticated owner via JWT
    - Soft-deleted resources (`is_active=false`) are treated as not found (404)
    - All monetary values are numbers (`NUMERIC(10,2)` precision)
  version: "1.0.0"
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # --------------------------------------------------------------
  # Reusable path parameters
  # --------------------------------------------------------------
  parameters:
    page:
      name: page
      in: query
      description: Page number (1-based)
      schema:
        type: integer
        default: 1
        minimum: 1
    page_size:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
    due_public_id:
      name: due_public_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    expense_public_id:
      name: expense_public_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    category_public_id:
      name: category_public_id
      in: path
      required: true
      schema:
        type: string
        format: uuid

  # --------------------------------------------------------------
  # Reusable error responses
  # --------------------------------------------------------------
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Authentication required"
    Forbidden:
      description: Authenticated but not authorized for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Access denied"
    NotFound:
      description: Resource not found or soft-deleted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            data: null
            message: "Not found"
    BadRequest:
      description: Business logic error (e.g. duplicate unit, conflicting dates, default category delete)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorEnvelope'

  # --------------------------------------------------------------
  # Schemas
  # --------------------------------------------------------------
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
          example: null
        message:
          type: string
    ValidationErrorEnvelope:
      type: object
      required: [success, data, message]
      properties:
        success:
          type: boolean
          example: false
        data:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
        message:
          type: string
          example: "Validation failed"
    Pagination:
      type: object
      required: [total, page, page_size]
      properties:
        total:
          type: integer
          description: Total number of matching records
        page:
          type: integer
        page_size:
          type: integer

    # ── Auth ───────────────────────────────────────────────────
...
    MonthlyDueData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        tenant_public_id:
          type: string
          format: uuid
        agreement_public_id:
          type: string
          format: uuid
        month:
          type: integer
          minimum: 1
          maximum: 12
        year:
          type: integer
        rent_amount:
          type: number
        total_due:
          type: number
        amount_paid:
          type: number
        remaining_balance:
          type: number
        status:
          type: string
          enum: [unpaid, partial, paid]
        is_auto_generated:
          type: boolean
        due_date:
          type: string
          format: date
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ── Payments ───────────────────────────────────────────────
    PaymentCreate:
      type: object
      required: [amount_paid, paid_on]
      properties:
        amount_paid:
          type: number
          minimum: 0
          exclusiveMinimum: true
        paid_on:
          type: string
          format: date
        note:
          type: string
          nullable: true
...
    PaymentRecordData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        due_public_id:
          type: string
          format: uuid
        amount_paid:
          type: number
        paid_on:
          type: string
          format: date
        note:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...
    ExpenseCreate:
      type: object
      required: [category_public_id, amount, expense_date, scope]
      properties:
        building_public_id:
          type: string
          format: uuid
          nullable: true
          description: "Required when scope=building or scope=apartment"
        apartment_public_id:
          type: string
          format: uuid
          nullable: true
          description: "Required when scope=apartment; building_public_id must also be set"
        category_public_id:
          type: string
          format: uuid
        description:
          type: string
          nullable: true
        amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
        expense_date:
          type: string
          format: date
        scope:
          type: string
          enum: [building, apartment]
        is_tenant_charged:
          type: boolean
          default: false
        due_public_id:
          type: string
          format: uuid
          nullable: true
          description: "Required when is_tenant_charged=true"
        charged_amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
          nullable: true
          description: "Required when is_tenant_charged=true; amount charged to the tenant's due"
    ExpenseUpdate:
      type: object
      properties:
        category_public_id:
          type: string
          format: uuid
        description:
          type: string
          nullable: true
        amount:
          type: number
          minimum: 0
          exclusiveMinimum: true
        expense_date:
          type: string
          format: date
        is_tenant_charged:
          type: boolean
    ExpenseData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        owner_public_id:
          type: string
          format: uuid
        building_public_id:
          type: string
          format: uuid
          nullable: true
        apartment_public_id:
          type: string
          format: uuid
          nullable: true
        category_public_id:
          type: string
          format: uuid
        category_name:
          type: string
        description:
          type: string
          nullable: true
        amount:
          type: number
        expense_date:
          type: string
          format: date
        scope:
          type: string
          enum: [building, apartment]
        is_tenant_charged:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ── Expense Categories ─────────────────────────────────────
    CategoryCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
        is_default:
          type: boolean
          default: false
    CategoryUpdate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
    CategoryData:
      type: object
      properties:
        public_id:
          type: string
          format: uuid
        name:
          type: string
        is_default:
          type: boolean
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time

    # ── Dashboard ──────────────────────────────────────────────
...
paths:
...
  /dues/{due_public_id}/payments:
    parameters:
      - $ref: '#/components/parameters/due_public_id'
    get:
      tags: [Payments]
      summary: List all payment records for a monthly due
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
      responses:
        "200":
          description: Payment records
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/PaymentRecordData'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Payments]
      summary: Record a payment against a monthly due
      description: |
        Inserts a PaymentRecord and recalculates remaining_balance on the MonthlyDue.
        Status transitions: unpaid → partial → paid (never backwards).
        Returns 400 if amount_paid would exceed remaining_balance.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentCreate'
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      payment:
                        $ref: '#/components/schemas/PaymentRecordData'
                      due:
                        $ref: '#/components/schemas/MonthlyDueData'
                  message: { type: string, example: "Payment recorded" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
...
  /expenses:
    get:
      tags: [Expenses]
      summary: List expenses for the authenticated owner
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
        - name: scope
          in: query
          schema:
            type: string
            enum: [building, apartment]
        - name: building_public_id
          in: query
          schema:
            type: string
            format: uuid
        - name: apartment_public_id
          in: query
          schema:
            type: string
            format: uuid
        - name: category_public_id
          in: query
          schema:
            type: string
            format: uuid
        - name: is_tenant_charged
          in: query
          schema:
            type: boolean
        - name: from_date
          in: query
          schema:
            type: string
            format: date
        - name: to_date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Expenses list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/ExpenseData'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Expenses]
      summary: Create an expense
      description: |
        When is_tenant_charged=true, must also provide due_public_id and charged_amount.
        A due_expense record is created atomically.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreate'
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/ExpenseData'
                  message: { type: string, example: "Expense created" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
  /expenses/{expense_public_id}:
    parameters:
      - $ref: '#/components/parameters/expense_public_id'
    put:
      tags: [Expenses]
      summary: Update an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseUpdate'
      responses:
        "200":
          description: Expense updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/ExpenseData'
                  message: { type: string, example: "Expense updated" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Expenses]
      summary: Soft delete an expense
      responses:
        "200":
          description: Expense deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { nullable: true, example: null }
                  message: { type: string, example: "Expense deactivated" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

  # ==============================================================
  # EXPENSE CATEGORIES
  # ==============================================================
  /expense-categories:
    get:
      tags: [Expense Categories]
      summary: List expense categories for the authenticated owner
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/page_size'
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: '#/components/schemas/CategoryData'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
                  message: { type: string, example: "OK" }
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Expense Categories]
      summary: Create an expense category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryCreate'
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/CategoryData'
                  message: { type: string, example: "Category created" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'
  /expense-categories/{category_public_id}:
    parameters:
      - $ref: '#/components/parameters/category_public_id'
    put:
      tags: [Expense Categories]
      summary: Update a category name
      description: "Only name is editable. is_default cannot be changed after creation."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpdate'
      responses:
        "200":
          description: Category updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/CategoryData'
                  message: { type: string, example: "Category updated" }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: '#/components/responses/UnprocessableEntity'

    delete:
      tags: [Expense Categories]
      summary: Soft delete a category
      description: "Returns 400 if is_default=true or if the category has active expenses."
      responses:
        "200":
          description: Category deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { nullable: true, example: null }
                  message: { type: string, example: "Category deactivated" }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

  # ==============================================================
  # DASHBOARD
  # ==============================================================
...